feat: Add admin console and database viewer with beautiful UI
- Fix type mismatch issues (i32 → i64) in portfolio and links models
- Add database API endpoints for viewing tables, schemas, and executing SQL queries
- Add server console API for executing commands, viewing system info, processes, and logs
- Create beautiful database viewer component with:
  - Table browser with row counts
  - Column filtering with multiple operators
  - Sorting and pagination
  - SQL query editor with syntax highlighting
  - Real-time query execution with timing
  - Cell value preview and copy functionality
- Create server console component with:
  - Terminal-style interface with dark theme
  - System stats dashboard (CPU, RAM, Disk, Uptime)
  - Quick command buttons for common operations
  - Command history with navigation
  - Process monitor with real-time updates
  - Log viewer with service filtering
- Update admin dashboard with new Console and Database tabs
- Add smooth animations throughout the UI

+++ server/src/portfolio/models.rs
@@ -53,7 +53,7 @@ pub struct PortfolioCaseImage {
     pub id: i64,
     pub case_id: i64,
     pub image_url: String,
-    pub order_index: i32,
+    pub order_index: i64,
 }
 
 // Request DTOs
